Revert " DCAR: Add ability to follow MyGuardian tags from Article webviews in native apps "
