+++ dotcom-rendering/src/components/FollowButton.stories.tsx
@@ -1,0 +1,20 @@
+import { splitTheme } from '../../.storybook/decorators/splitThemeDecorator';
+import { FollowButton } from './FollowButton';
+
+export default {
+	component: FollowButton,
+	title: 'Components/FollowStatus',
+};
+
+export const FollowBothStates = () => {
+	return (
+		<>
+			<FollowButton
+				isFollowing={false}
+				onClickHandler={() => undefined}
+			/>
+			<FollowButton isFollowing={true} onClickHandler={() => undefined} />
+		</>
+	);
+};
+FollowBothStates.decorators = [splitTheme()];
