+++ dotcom-rendering/src/components/FollowWrapper.importable.tsx
@@ -1,11 +1,8 @@
 import { css } from '@emotion/react';
 import { Topic } from '@guardian/bridget/Topic';
-import { log } from '@guardian/libs';
 import { useEffect, useState } from 'react';
-import { getNotificationsClient, getTagClient } from '../lib/bridgetApi';
-import { useIsBridgetCompatible } from '../lib/useIsBridgetCompatible';
-import { useIsMyGuardianEnabled } from '../lib/useIsMyGuardianEnabled';
-import { FollowNotificationsButton, FollowTagButton } from './FollowButtons';
+import { getNotificationsClient } from '../lib/bridgetApi';
+import { FollowButton } from './FollowButton';
 
 type Props = {
 	id: string;
@@ -13,19 +10,9 @@ type Props = {
 };
 
 export const FollowWrapper = ({ id, displayName }: Props) => {
-	const [isFollowingNotifications, setIsFollowingNotifications] = useState<
-		boolean | undefined
-	>(undefined);
-	const [isFollowingTag, setIsFollowingTag] = useState<boolean | undefined>(
+	const [isFollowing, setIsFollowing] = useState<boolean | undefined>(
 		undefined,
 	);
-	const [showFollowTagButton, setShowFollowTagButton] =
-		useState<boolean>(false);
-
-	const isMyGuardianEnabled = useIsMyGuardianEnabled();
-	const isBridgetCompatible = useIsBridgetCompatible('2.5.0');
-	isBridgetCompatible && isMyGuardianEnabled && setShowFollowTagButton(true);
-
 	useEffect(() => {
 		const topic = new Topic({
 			id,
@@ -35,102 +22,37 @@ export const FollowWrapper = ({ id, displayName }: Props) => {
 
 		void getNotificationsClient()
 			.isFollowing(topic)
-			.then(setIsFollowingNotifications)
-			.catch((e) =>
-				log(
-					'dotcom',
-					'Bridget getNotificationsClient.isFollowing Error:',
-					e,
-				),
-			);
-
-		void getTagClient()
-			.isFollowing(topic)
-			.then(setIsFollowingTag)
-			.catch((e) =>
-				log('dotcom', 'Bridget getTagClient.isFollowing Error:', e),
-			);
+			.then(setIsFollowing)
+			.catch((e) => console.error(e));
 	}, [id, displayName]);
 
-	const tagHandler = () => {
+	const handler = () => {
 		const topic = new Topic({
 			id,
 			displayName,
 			type: 'tag-contributor',
 		});
 
-		isFollowingTag
-			? void getTagClient()
-					.unfollow(topic)
-					.then(() => setIsFollowingTag(false))
-					.catch((e) =>
-						log(
-							'dotcom',
-							'Bridget getTagClient.unfollow Error:',
-							e,
-						),
-					)
-			: void getTagClient()
-					.follow(topic)
-					.then(() => setIsFollowingTag(true))
-					.catch((e) =>
-						log('dotcom', 'Bridget getTagClient.follow Error:', e),
-					);
-	};
-
-	const notificationsHandler = () => {
-		const topic = new Topic({
-			id,
-			displayName,
-			type: 'tag-contributor',
-		});
-
-		isFollowingNotifications
+		isFollowing
 			? void getNotificationsClient()
 					.unfollow(topic)
-					.then(() => setIsFollowingNotifications(false))
-					.catch((e) =>
-						log(
-							'dotcom',
-							'Bridget getNotificationsClient.unfollow Error:',
-							e,
-						),
-					)
+					.then(() => setIsFollowing(false))
+					.catch((e) => console.error(e))
 			: void getNotificationsClient()
 					.follow(topic)
-					.then(() => setIsFollowingNotifications(true))
-					.catch((e) =>
-						log(
-							'dotcom',
-							'Bridget getNotificationsClient.follow Error:',
-							e,
-						),
-					);
+					.then(() => setIsFollowing(true))
+					.catch((e) => console.error(e));
 	};
-
 	return (
 		<div
 			css={css`
 				min-height: 24px;
 			`}
 		>
-			{showFollowTagButton && (
-				<FollowTagButton
-					isFollowing={isFollowingTag ?? false}
-					displayName={displayName}
-					onClickHandler={
-						isFollowingTag !== undefined
-							? tagHandler
-							: () => undefined
-					}
-				/>
-			)}
-			<FollowNotificationsButton
-				isFollowing={isFollowingNotifications ?? false}
+			<FollowButton
+				isFollowing={isFollowing ?? false}
 				onClickHandler={
-					isFollowingNotifications !== undefined
-						? notificationsHandler
-						: () => undefined
+					isFollowing !== undefined ? handler : () => undefined
 				}
 			/>
 		</div>
