+++ dotcom-rendering/src/lib/bridgetApi.ts
@@ -7,7 +7,6 @@ import * as Metrics from '@guardian/bridget/Metrics';
 import * as Navigation from '@guardian/bridget/Navigation';
 import * as Newsletters from '@guardian/bridget/Newsletters';
 import * as Notifications from '@guardian/bridget/Notifications';
-import * as Tag from '@guardian/bridget/Tag';
 import * as User from '@guardian/bridget/User';
 import * as Video from '@guardian/bridget/Videos';
 import { createAppClient } from './thrift/nativeConnection';
@@ -58,18 +57,6 @@ export const getNotificationsClient = (): Notifications.Client<void> => {
 		);
 	}
 	return notificationsClient;
-};
-
-let tagClient: Tag.Client<void> | undefined = undefined;
-export const getTagClient = (): Tag.Client<void> => {
-	if (!tagClient) {
-		tagClient = createAppClient<Tag.Client<void>>(
-			Tag.Client,
-			'buffered',
-			'compact',
-		);
-	}
-	return tagClient;
 };
 
 let userClient: User.Client<void> | undefined = undefined;
